--- final_project/plift.java
@@ -1,6 +1,0 @@
-public class plift {
-    //Irina is trying to figure github out
-
-    //egor did some work too
-
-}
