+++ final_project/plift/Main.java
@@ -1,4 +1,4 @@
-package test;
+package plift;
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
@@ -10,8 +10,8 @@ public class Main extends Application {
 
     @Override
     public void start(Stage primaryStage) throws Exception{
-        Parent root = FXMLLoader.load(getClass().getResource("test.fxml"));
-        primaryStage.setTitle("Hello World");
+        Parent root = FXMLLoader.load(getClass().getResource("Splash.fxml"));
+        primaryStage.setTitle("Welcome to Lifting Calculator!");
         primaryStage.setScene(new Scene(root, 300, 275));
         primaryStage.show();
     }
