+++ final_project/plift/Splash.fxml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.HBox?>
+
+<AnchorPane id="splash-background" prefHeight="275.0" prefWidth="300" stylesheets="@style.css" xmlns="http://javafx.com/javafx/8.0.172-ea" xmlns:fx="http://javafx.com/fxml/1" fx:controller="plift.Controller">
+   <children>
+      <HBox prefHeight="200.0" prefWidth="300.0" />
+      <Button alignment="CENTER" layoutX="106.0" layoutY="200.0" mnemonicParsing="false" prefHeight="47.0" prefWidth="88.0" text="Start!" />
+   </children>
+</AnchorPane>
