+++ final_project/plift/plift_weightplate.java
@@ -1,2 +1,4 @@
+package plift;
+
 public class plift_weightplate {
 }
