+++ final_project/plift/style.css
@@ -1,0 +1,3 @@
+#splash-background{
+    -fx-background-image : url("splash.jpg")
+}
