--- final_project/test/test.fxml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.CheckBox?>
-<?import javafx.scene.control.ComboBox?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.layout.HBox?>
-
-
-<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.172-ea" xmlns:fx="http://javafx.com/fxml/1" fx:controller="test.Controller">
-   <children>
-      <HBox layoutX="17.0" layoutY="135.0" prefHeight="100.0" prefWidth="575.0">
-         <children>
-            <Button mnemonicParsing="false" text="Button" />
-            <CheckBox mnemonicParsing="false" text="CheckBox" />
-            <ComboBox prefWidth="150.0" />
-         </children>
-      </HBox>
-   </children>
-</AnchorPane>
